+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Add `fetchMetadataAccessCreds` function option in `ToprfSecureBackup` class constructor.
   - This function should be used to inject metadata access token for calling metadata apis.
 
+### Changed
+
+- Changed timestamp format from seconds to milliseconds in reset rate limit requests to minimize nonce collision. ([#133](https://github.com/MetaMask/toprf-secure-backup/pull/133))
+
 ## [0.5.0]
 
 ### Added