+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **Breaking**: Add `fetchMetadataAccessCreds` function option in `ToprfSecureBackup` class constructor.
+- **Breaking**: Add `fetchMetadataAccessCreds` function option in `ToprfSecureBackup` class constructor.([#134]https://github.com/MetaMask/toprf-secure-backup/pull/134)
   - This function should be used to inject metadata access token for calling metadata apis.
 
 ### Changed